+++ app/admin/content/components/SocialTab.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useState, useRef } from 'react'
+import { useState } from 'react'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { Drawer } from '@/components/ui/drawer'
 import {
@@ -18,7 +18,8 @@ import { AdminStatusBadge } from '@/app/admin/components/AdminStatusBadge'
 import { useAdminDrawer } from '@/app/admin/components/useAdminDrawer'
 import { makeDragHandlers } from './useDragSort'
 import { useLanguage } from '@/lib/hooks/useLanguage'
-import { isCdnUrl, isStorageUrl } from '@/lib/social-thumbnail'
+import { SocialPostForm, InstagramIcon, FacebookIcon } from './SocialPostForm'
+import type { SocialPostFormData } from './SocialPostForm'
 
 type SocialPost = {
   id: string
@@ -31,216 +32,6 @@ type SocialPost = {
   sort_order: number
   created_at: string
   posted_at: string | null
-}
-
-type SocialPostFormData = {
-  platform: 'instagram' | 'facebook'
-  post_url: string
-  caption: string
-  thumbnail_url: string
-  posted_at: string
-}
-
-function InstagramIcon() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 24 24" fill="none">
-      <rect x="2" y="2" width="20" height="20" rx="5" stroke="currentColor" strokeWidth="1.8"/>
-      <circle cx="12" cy="12" r="4" stroke="currentColor" strokeWidth="1.8"/>
-      <circle cx="17.5" cy="6.5" r="1" fill="currentColor"/>
-    </svg>
-  )
-}
-
-function FacebookIcon() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 24 24" fill="none">
-      <path d="M18 2h-3a5 5 0 0 0-5 5v3H7v4h3v8h4v-8h3l1-4h-4V7a1 1 0 0 1 1-1h3z" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"/>
-    </svg>
-  )
-}
-
-function SocialPostForm({
-  initial,
-  onSave,
-  onCancel,
-  isPending,
-  error,
-}: {
-  initial: SocialPostFormData
-  onSave: (data: SocialPostFormData) => void
-  onCancel: () => void
-  isPending: boolean
-  error: string | null
-}) {
-  const { t } = useLanguage()
-  const [form, setForm] = useState(initial)
-  const [previewing, setPreviewing] = useState(false)
-  const [previewHint, setPreviewHint] = useState<string | null>(null)
-  const [uploading, setUploading] = useState(false)
-  const fileInputRef = useRef<HTMLInputElement>(null)
-
-  const thumbnailIsCdn = form.thumbnail_url ? isCdnUrl(form.thumbnail_url) : false
-  const thumbnailIsStorage = form.thumbnail_url ? isStorageUrl(form.thumbnail_url) : false
-  const thumbnailIsBlocked = thumbnailIsCdn && !thumbnailIsStorage
-
-  async function fetchOgPreview(url: string) {
-    if (!url) return
-    setPreviewing(true)
-    setPreviewHint(null)
-    try {
-      const res = await fetch(`/api/admin/social-posts/preview?url=${encodeURIComponent(url)}`)
-      if (!res.ok) throw new Error('preview failed')
-      const data = await res.json() as { thumbnail_url: string | null; caption: string | null }
-      setForm(f => ({
-        ...f,
-        thumbnail_url: data.thumbnail_url ?? f.thumbnail_url,
-        caption: data.caption ?? f.caption,
-      }))
-      if (!data.thumbnail_url && !data.caption) {
-        setPreviewHint(t('admin.content.social.previewUnavailable'))
-      }
-    } catch {
-      setPreviewHint(t('admin.content.social.previewUnavailable'))
-    } finally {
-      setPreviewing(false)
-    }
-  }
-
-  async function handleFileUpload(file: File) {
-    setUploading(true)
-    try {
-      const formData = new FormData()
-      formData.append('file', file)
-      const res = await fetch('/api/admin/social-posts/upload-thumbnail', {
-        method: 'POST',
-        body: formData,
-      })
-      if (!res.ok) throw new Error('Upload failed')
-      const { url } = await res.json() as { url: string }
-      setForm(f => ({ ...f, thumbnail_url: url }))
-    } catch {
-      setPreviewHint(t('admin.content.social.previewUnavailable'))
-    } finally {
-      setUploading(false)
-    }
-  }
-
-  function handleSave() {
-    onSave(form)
-  }
-
-  return (
-    <div className="space-y-4">
-      <div className="flex gap-2">
-        {(['instagram', 'facebook'] as const).map(p => (
-          <button key={p} onClick={() => setForm(f => ({ ...f, platform: p }))}
-            className="flex items-center gap-2 px-4 py-2 rounded-xl text-sm font-semibold transition-all"
-            style={{ backgroundColor: form.platform === p ? 'var(--brand-forest)' : 'rgba(0,0,0,0.06)', color: form.platform === p ? 'var(--brand-parchment)' : 'var(--text-secondary)' }}>
-            {p === 'instagram' ? <InstagramIcon /> : <FacebookIcon />}
-            {p.charAt(0).toUpperCase() + p.slice(1)}
-          </button>
-        ))}
-      </div>
-      <div>
-        <input
-          value={form.post_url}
-          onChange={e => setForm(f => ({ ...f, post_url: e.target.value }))}
-          onBlur={e => { if (e.target.value) fetchOgPreview(e.target.value) }}
-          placeholder={t('admin.content.social.placeholder.postUrl')}
-          className="w-full border rounded-xl px-3 py-2.5 text-sm"
-          style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }}
-        />
-        {previewing && <p className="text-xs mt-1" style={{ color: 'var(--text-secondary)' }}>{t('admin.content.social.fetchingPreview')}</p>}
-        {!previewing && previewHint && <p className="text-xs mt-1" style={{ color: 'var(--text-secondary)' }}>{previewHint}</p>}
-      </div>
-      <textarea
-        value={form.caption}
-        onChange={e => setForm(f => ({ ...f, caption: e.target.value }))}
-        placeholder={t('admin.content.social.placeholder.caption')}
-        rows={3}
-        className="w-full border rounded-xl px-3 py-2.5 text-sm resize-none"
-        style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }}
-      />
-
-      <div>
-        <label className="block text-xs font-medium mb-1" style={{ color: 'var(--text-secondary)' }}>
-          Post date (optional)
-        </label>
-        <input
-          type="datetime-local"
-          value={form.posted_at}
-          onChange={e => setForm(f => ({ ...f, posted_at: e.target.value }))}
-          className="w-full border rounded-xl px-3 py-2.5 text-sm"
-          style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }}
-        />
-        <p className="text-xs mt-1" style={{ color: 'var(--text-secondary)', opacity: 0.7 }}>
-          Actual published date. Shown on the bento tile. Falls back to creation time if left empty.
-        </p>
-      </div>
-
-      <div className="space-y-2">
-        <div className="flex items-center gap-2">
-          <button
-            type="button"
-            onClick={() => fileInputRef.current?.click()}
-            disabled={uploading}
-            className="px-3 py-1.5 rounded-xl text-xs font-semibold border transition-colors hover:bg-black/5 disabled:opacity-50"
-            style={{ borderColor: 'var(--border-default)', color: 'var(--text-secondary)' }}
-          >
-            {uploading ? t('admin.content.social.btn.uploading') : t('admin.content.social.btn.uploadThumbnail')}
-          </button>
-          {form.thumbnail_url && thumbnailIsStorage && (
-            <span className="text-xs" style={{ color: 'var(--text-secondary)' }}>
-              ✓ {t('admin.content.social.thumbnailStored')}
-            </span>
-          )}
-          <input
-            ref={fileInputRef}
-            type="file"
-            accept="image/jpeg,image/png,image/webp,image/gif"
-            className="hidden"
-            onChange={e => {
-              const file = e.target.files?.[0]
-              if (file) handleFileUpload(file)
-              e.target.value = ''
-            }}
-          />
-        </div>
-        <input
-          value={form.thumbnail_url}
-          onChange={e => setForm(f => ({ ...f, thumbnail_url: e.target.value }))}
-          placeholder={t('admin.content.social.placeholder.thumbnailUrl')}
-          className="w-full border rounded-xl px-3 py-2.5 text-sm"
-          style={{
-            borderColor: thumbnailIsBlocked ? 'var(--brand-crimson)' : 'var(--border-default)',
-            color: 'var(--text-primary)',
-            backgroundColor: 'var(--bg-card)',
-          }}
-        />
-        {thumbnailIsBlocked && (
-          <p className="text-xs" style={{ color: 'var(--brand-crimson)' }}>
-            {t('admin.content.social.cdnUrlWarning')}
-          </p>
-        )}
-      </div>
-
-      {error && <p className="text-xs" style={{ color: 'var(--brand-crimson)' }}>{error}</p>}
-      <div className="flex gap-3 pt-2">
-        <button
-          onClick={handleSave}
-          disabled={isPending || !form.post_url || thumbnailIsBlocked}
-          className="px-6 py-2.5 rounded-xl text-sm font-semibold text-white disabled:opacity-40 transition-opacity hover:opacity-90"
-          style={{ backgroundColor: 'var(--brand-crimson)' }}>
-          {isPending ? t('admin.content.social.btn.saving') : t('admin.content.social.btn.save')}
-        </button>
-        <button onClick={onCancel}
-          className="px-6 py-2.5 rounded-xl text-sm font-semibold border transition-colors hover:bg-black/5"
-          style={{ borderColor: 'var(--border-default)', color: 'var(--text-secondary)' }}>
-          {t('admin.content.social.btn.cancel')}
-        </button>
-      </div>
-    </div>
-  )
 }
 
 export function SocialTab() {
